Support manual ID in CliManualDeleter
It's possible (due to a bug) to create multiple manuals with the same
slug. The CliManualDeleter correctly gives an ambiguous slug error
and refuses to delete anything.

This commit allows `bin/delete_draft_manual` to be called with an
optional identifier in addition to the slug. If the identifier is
present it will be used instead of the slug.